backend-flask/pyproject.toml
[project]
name = "ttps-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Leandro Daniel Romagnoli",email = "romagnolileandro02@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "pypdf (>=6.1.1,<7.0.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymupdf (>=1.26.4,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "twilio (>=8.0.0)",
    "httpx (>=0.24.0)",
    "openai (==0.28.0)",
    "phonenumbers (>=9.0.19,<10.0.0)",
    "flask-sqlalchemy-lite (>=0.2.1,<0.3.0)",
    "apscheduler (>=3.10.0,<4.0.0)",
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ttps-chatbot"
version = "0.1.0"
package-mode = false